requirements.txt
fastapi
uvicorn[standard]
pandas
python-multipart
scikit-learn==1.4.0
numpy==1.26.3
openai>=1.30.0
python-dotenv==1.0.0
SQLAlchemy
reportlab
XlsxWriter